:recycle: api: move CertificationIssueReportCategory read model to shared
